requirements.txt
pip==23.0.1
setuptools==67.6.0
Flask==2.2.3
Flask-Argon2==0.3.0.0
mysqlclient==2.1.1
Jinja2==3.1.2
MarkupSafe==2.1.2
WTForms==3.0.1
Werkzeug==2.2.3
argon2-cffi-bindings==21.2.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
idna==3.4
itsdangerous==2.1.2
pyargon2==1.0.7
pycparser==2.21
pyserial==3.5
six==1.16.0
urllib3==1.26.15
Flask-Login==0.6.2
sms==0.4
Flask-MySQLdb==1.0.1
requests==2.28.2
Flask-WTF==1.1.1
argon2-cffi==21.3.0
argon2pure==1.3
pyOpenSSL==23.1.1